Changed float position to float64

+++ cardinal/component/position.go
@@ -1,7 +1,7 @@
 package component
 
 type Position struct {
-	Island    [2]int     `json:"island"`
+	Island    [2]float64 `json:"island"`
 	Shipwreck [2]float64 `json:"shipwreck"`
 }
